requirements.txt
ultralytics>=8.2.0
opencv-python>=4.8.0
numpy>=1.24
pandas>=2.0
PyYAML>=6.0
tqdm>=4.66